starting to use new tactics in Engine

+++ truffle-core/contracts/Leagues.sol
@@ -94,7 +94,7 @@ contract Leagues is Assets {
     function computeMatchday(
         uint8 matchday,
         uint256[PLAYERS_PER_TEAM_MAX][TEAMS_PER_LEAGUE] memory prevLeagueState,
-        uint8[TEAMS_PER_LEAGUE] memory tacticsIds,
+        uint256[TEAMS_PER_LEAGUE] memory tacticsIds,
         uint256 currentVerseSeed
     )
         public
@@ -109,7 +109,7 @@ contract Leagues is Assets {
         {
             (homeTeamIdx, visitorTeamIdx) = getTeamsInMatch(matchday, matchIdxInDay);
             uint256 matchSeed = uint256(keccak256(abi.encode(currentVerseSeed, matchIdxInDay))); 
-            uint8[2] memory tactics = [tacticsIds[homeTeamIdx], tacticsIds[visitorTeamIdx]];
+            uint256[2] memory tactics = [tacticsIds[homeTeamIdx], tacticsIds[visitorTeamIdx]];
             score = _engine.playMatch(
                 matchSeed, 
                 prevLeagueState[homeTeamIdx],
